Rename ProxyBackend.h to ProxyBackend.hpp
